+++ run.py
@@ -5,7 +5,7 @@ import os
 from dotenv import load_dotenv
 import pandas as pd
 
-from modules.extract import download_dataset
+from modules.etl import download_dataset
 from modules.data_processing import (prepare_dataframe_tris, prepare_dataframe_melate_retro, filter_dataframe_by_year,
                                      get_probability_of_numbers_per_column, get_probability_of_numbers_in_all_columns,
                                      plot_probabilities, count_winning_numbers)
